tests: Remove existing tests
They depended on headers from the deprecated and archived boost-dbus
repo, which in turn depends on ancient boost releases. Other
repositories in the openbmc github organisation depend on APIs only
available in newer boost releases, and so we're wedged.

This is all pretty terrible, so allow new tests to be written by
dropping the existing code.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Change-Id: I95b28d214d7a4708f4427afe4a2c77e19409cc1f